Big overhaul to env/game, add utils
Change what each is responsible for.

Game tracks the relevant info for all the entities in the game.
It is responsible for updating these entities given an action, and
is also responsible for generating new games.

Environment is responsible for pretty much everything else:
maintaining the display of the current game, the last game summary,
determining what the action and observation spaces should be, etc.

The environment is now configured with a single argument "env_config"
which holds everything necessary to create it. This is in accordance
with the RLLib library, which may get used in our plug'n'play efforts.